refactor: create `PendingFlows` abstraction
